style: rename scappers to scrappers
